pyproject.toml
# pyproject.toml - PEP 621 compliant package metadata
# Document Consolidation Toolkit

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-consolidation-toolkit"
version = "0.1.0"
description = "Tournament-based document consolidation for legal practice"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "LROC", email = "contact@lroc.com"}
]
maintainers = [
    {name = "LROC", email = "contact@lroc.com"}
]
keywords = [
    "legal",
    "document-consolidation",
    "document-management",
    "tournament-ranking",
    "legal-tech",
    "markdown",
    "legal-documents"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Office/Business",
    "Typing :: Typed",
]

# Core dependencies - production requirements
dependencies = [
    "pyyaml>=6.0",           # YAML configuration file parsing
    "pydantic>=2.0,<3.0",    # Settings validation and data models
    "click>=8.0,<9.0",       # CLI framework
    "tqdm>=4.65.0",          # Progress bars
]

# Optional dependency groups
[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",

    # Code quality
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",

    # Type stubs
    "types-PyYAML>=6.0",

    # Development tools
    "ipython>=8.0",
    "ipdb>=0.13",
]

[project.urls]
Homepage = "https://github.com/LROC-NY/document-consolidation-toolkit"
Repository = "https://github.com/LROC-NY/document-consolidation-toolkit"
Issues = "https://github.com/LROC-NY/document-consolidation-toolkit/issues"
Documentation = "https://github.com/LROC-NY/document-consolidation-toolkit/blob/main/README.md"

# CLI entry points
[project.scripts]
consolidate = "document_consolidation.cli:main"

# Setuptools configuration
[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["document_consolidation*"]
namespaces = false

[tool.setuptools.package-data]
document_consolidation = ["py.typed"]  # PEP 561 marker for type information

# Black configuration
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py39"

# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "EM",  # flake8-errmsg
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "PD",  # pandas-vet
    "PL",  # pylint
    "RUF", # Ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",    # line too long (handled by Black)
    "PLR0913", # too many arguments to function call
    "PLR2004", # magic value used in comparison
]

# Exclude directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Per-file ignores
[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["ARG001", "PLR2004"]  # Allow unused args and magic values in tests

[tool.ruff.isort]
known-first-party = ["document_consolidation"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Test discovery patterns
addopts = [
    "-ra",                    # Show extra test summary info for all outcomes
    "--strict-markers",       # Treat unregistered markers as errors
    "--strict-config",        # Treat config errors as errors
    "-p", "no:cacheprovider", # Disable pytest cache
]

# Custom markers
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, uses filesystem)",
    "slow: Slow tests (mark for optional execution)",
]

# Logging during tests
log_cli = false
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# Coverage options (when using pytest-cov)
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstract",
]

[tool.coverage.html]
directory = "htmlcov"

# Mypy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict_concatenate = true

# Strict mode (equivalent to --strict)
strict = true

# Output configuration
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true

# Per-module options
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false  # Allow untyped test functions

[[tool.mypy.overrides]]
module = "tqdm.*"
ignore_missing_imports = true  # tqdm doesn't have complete type stubs

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true  # PyYAML type stubs handled by types-PyYAML
